requirements.txt
streamlit==1.25.0
flask==2.3.3
requests==2.31.0
pdfplumber==0.10.2
python-dotenv==1.0.0
langchain==0.0.230
openai==0.28.0
logging==0.5.1.2
